Rename module owwidget to concurrent
