+++ orangecontrib/text/widgets/utils/__init__.py
@@ -1,1 +1,2 @@
 from .widgets import *
+from .concurrent import OWConcurrentWidget, asynchronous
