+++ orangecontrib/text/widgets/utils/concurrent.py
